+++ client/src/games/types/Popularity.js
@@ -1,0 +1,5 @@
+import PropTypes from 'prop-types';
+
+export const PopularityOptionsPropType = PropTypes.shape({
+    vertical: PropTypes.bool
+});